requirements.txt
blinker==1.9.0
cachetools==5.5.1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
Flask==3.1.0
google-auth==2.38.0
google-auth-oauthlib==1.2.1
gspread==6.1.4
gunicorn==23.0.0
httplib2==0.22.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
mysql-connector-python==9.2.0
oauth2client==4.1.3
oauthlib==3.2.2
packaging==24.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyparsing==3.2.1
python-dotenv==1.0.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
six==1.17.0
urllib3==2.3.0
Werkzeug==3.1.3
